refactored majority of files. purge of unnecessary code and reorganization
