+++ frontend/src/PrivateRoute.js
@@ -3,7 +3,6 @@ import {Route, Redirect } from 'react-router-dom';
 import Storage from './storage'
 
 const PrivateRoute = ({ component: Component, ...rest }) => {
-  console.log(Storage.isAuthenticated())
 return (
   <Route {...rest} render={(props) => (
     Storage.isAuthenticated()
