+++ frontend/src/login/login.css
@@ -2,34 +2,6 @@ body {
   background-color: #E7DFDD;
 }
 
-.img1 {
-  background-color: #849974;
-  margin-left: 12px;
-  margin-top: 45px;
-}
-
-.img2 {
-  background-color: #B2B2A2;
-  margin-left: 12px;
-  margin-top: 45px;
-}
-
-.img3 {
-  background-color: #613A43;
-  margin-left: 12px;
-  margin-top: 45px;
-}
-
-.logo {
-  /* background-image: url('../images/lined-paper.jpeg'); */
-  margin-left: 0%;
-  margin-top: 0%;
-  background-position: center;
-  position: absolute;
-  background-repeat: no-repeat;
-
-}
-
 .login {
   position: relative;
   background-color: #36384C;
@@ -42,10 +14,9 @@ body {
   left: 44%;
   top: 46%;
 }
+
 .Navbar {
   position: sticky;
-
-
 }
 
 .grid {
