+++ frontend/src/time/time.js
@@ -2,12 +2,11 @@ import React, { Component }  from 'react';
 import Clock from 'react-live-clock';
 import './Time.css'
 
-
 class Time extends Component {
   render() {
     return (
       <div>
-          <Clock className="time" format={"dddd, MMMM Do YYYY, h:mm a"} ticking={true} timezone={'US/Mountain'} />
+        <Clock className="time" format={"dddd, MMMM Do YYYY, h:mm a"} ticking={true} timezone={'US/Mountain'} />
       </div>
     );
   }
