+++ server.js
@@ -1,29 +1,27 @@
-const app = require('./app');
+const middleware = require('./middleware');
 const debug = require('debug')('backend:server');
 const http = require('http');
 const express = require('express');
 
 let port = normalizePort(process.env.PORT || '4001');
-app.set('port', port);
+middleware.set('port', port);
 if (process.env.NODE_ENV === "production") {
-  app.use(express.static("frontend/build"));
+  middleware.use(express.static("frontend/build"));
 }
 
-let server = http.createServer(app);
+let server = http.createServer(middleware);
 
 function normalizePort(val) {
     let port = parseInt(val, 10);
 
     if (isNaN(port)) {
-        // named pipe
-        return val;
+      // named pipe
+      return val;
     }
-
     if (port >= 0) {
-        // port number
-        return port;
+      // port number
+      return port;
     }
-
     return false;
 }
