+++ utils/token-utils.js
@@ -3,21 +3,21 @@ const config = require('../config/config')
 const { someOtherSecret } =  config.googleAuth
 
 const createToken = function(auth) {
-    return jwt.sign({
-            id: auth.id
-        }, someOtherSecret,
-        {
-            expiresIn: 60 * 120
-        });
+  return jwt.sign({
+      id: auth.id
+    }, someOtherSecret,
+    {
+      expiresIn: 60 * 120
+    });
 };
 
 module.exports = {
   generateToken: function(req, res, next) {
-      req.token = createToken(req.auth);
-      return next();
+    req.token = createToken(req.auth);
+    return next();
   },
   sendToken: function(req, res) {
-      res.setHeader('x-auth-token', req.token);
-      return res.status(200).send(JSON.stringify(req.user));
+    res.setHeader('x-auth-token', req.token);
+    return res.status(200).send(JSON.stringify(req.user));
   }
 };
